# swagger.yaml
openapi: 3.0.0
info:
  title: Formyy API
  version: 1.0.0
paths:
...
  /forms/statusChange:
    get:
      summary: Change form status
      tags:
        - Form
      parameters:
        - in: query
          name: id
          schema:
            type: string
          required: true
          description: The ID of the form
        - in: query
          name: state
          schema:
            type: string
          required: true
          description: The new state of the form ("active" or "inactive")
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid request
  /form/delete/{id}:
    get:
      summary: Delete form by ID
      tags:
        - Form
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
...
  /form/{id}:
    get:
      summary: Get form by ID
      tags:
        - Form
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
...
  /form/{id}/submissions/{sub_id}/delete:
    get:
      summary: Delete form submission by ID
      tags:
        - Form
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: sub_id
        required: true
        schema:
          type: string
          format: uuid